+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Load genes, transcripts and exons from pre-downloaded files
 - Demo coverage report endpoint in new `report` module
+- Coverage completeness lines in HTML coverage report
 ### Changed
 - Moved helper function from endpoints coverage to crud samples
 - Deleted unused `src/chanjo2/meta/handle_query_intervals.py` file